requirements.txt
absl-py==2.1.0
albucore @ file:///home/conda/feedstock_root/build_artifacts/albucore_1724942913978/work
albumentations @ file:///home/conda/feedstock_root/build_artifacts/albumentations_1724139891215/work
annotated-types @ file:///croot/annotated-types_1709542908624/work
Bottleneck @ file:///croot/bottleneck_1707864210935/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
contourpy @ file:///croot/contourpy_1700583582875/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
decorator==4.4.2
eval_type_backport @ file:///home/conda/feedstock_root/build_artifacts/eval_type_backport_1724150075856/work
fonttools @ file:///croot/fonttools_1713551344105/work
h5py==3.11.0
-e git+https://github.com/hhj1897/face_alignment.git@9cf5494e443f26d567972f3f50f6212d65b76c01#egg=ibug_face_alignment
-e git+https://github.com/hhj1897/face_detection.git@db2a4e8eae8c9c53385ff0773e9db08f03cf21ad#egg=ibug_face_detection
idna @ file:///croot/idna_1714398848350/work
imagecodecs @ file:///croot/imagecodecs_1695064943445/work
imageio @ file:///croot/imageio_1707247282708/work
imageio-ffmpeg==0.5.1
importlib_resources @ file:///croot/importlib_resources-suite_1720641103994/work
keras==3.5.0
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lazy_loader @ file:///croot/lazy_loader_1718176737906/work
markdown-it-py @ file:///croot/markdown-it-py_1684279902645/work
matplotlib==3.9.2
mdurl @ file:///opt/conda/conda-bld/mdurl_1659716024347/work
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
ml-dtypes==0.4.0
moviepy==1.0.3
mtcnn==0.1.1
namex==0.0.8
networkx @ file:///croot/networkx_1717597493534/work
numexpr @ file:///croot/numexpr_1696515281613/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp39-cp39-linux_x86_64.whl#sha256=6094eeedd869502faa0fd0a8c5ad3a70c5779be06ddd1feb7627e5c212fac420
opencv-python==4.10.0
opencv-python-headless==4.10.0
optree==0.12.1
packaging @ file:///croot/packaging_1720101850331/work
pandas @ file:///croot/pandas_1718308974269/work/dist/pandas-2.2.2-cp39-cp39-linux_x86_64.whl#sha256=484b9c988a0aaf85ef9256891a79639c95dfb3cacfd316d465b4f319dfef8d07
pillow @ file:///croot/pillow_1721059439630/work
ply==3.11
proglog==0.1.10
pydantic @ file:///croot/pydantic_1725040526370/work
pydantic_core @ file:///croot/pydantic-core_1724790363792/work
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing @ file:///croot/pyparsing_1725041624202/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
requests @ file:///croot/requests_1721410876868/work
rich @ file:///home/conda/feedstock_root/build_artifacts/rich-split_1709150387247/work/dist
scikit-image @ file:///croot/scikit-image_1707346116243/work
scipy @ file:///croot/scipy_1717521478074/work/dist/scipy-1.13.1-cp39-cp39-linux_x86_64.whl#sha256=2cb3d1436f8762a53161486cb9493ccc305fc760e8f830f82888792cbaf99099
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
tifffile @ file:///croot/tifffile_1695107451082/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado @ file:///croot/tornado_1718740109488/work
tqdm @ file:///croot/tqdm_1724853939799/work
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
urllib3 @ file:///croot/urllib3_1718912636303/work
zipp @ file:///croot/zipp_1704206909481/work
